Fix CustomerProfile form so that Formik does not unmount; Fix Button props to accept 'type';
